fix(lint): satisfy golangci-lint v2; move skip tests into router_test.go
- resolveSkip returns a named skipResult struct instead of three bare
  ints, resolving the gocritic unnamedResult finding without tripping
  nonamedreturns.
- Reorder the App struct so skipHasUseRoutes groups with the other
  trailing bool fields, fixing the govet fieldalignment finding.
- Move all SkipUnmatchedRoutes tests and benchmarks from
  router_skip_test.go into router_test.go, switch request bodies from nil
  to http.NoBody (gocritic httpNoBody), and flatten the group-middleware
  test so it no longer has order-dependent non-parallel subtests
  (tparallel).

golangci-lint v2.12.2 now reports 0 issues; full suite passes; new-code
coverage remains ~100%.
